refactor: Extract out PlaylistModal as component
