+++ src/components/VideoCard/VideoCard.jsx
@@ -1,7 +1,8 @@
 import "./videocard.css";
 import { useState } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { useAuth, useLikedList, useWatchLaterList, useHistoryList, usePlaylist } from "../../contexts";
+import { useAuth, useLikedList, useWatchLaterList, useHistoryList } from "../../contexts";
+import { PlaylistModal } from "..";
 
 const VideoCard = ({ video }) => {
 
@@ -17,8 +18,6 @@ const VideoCard = ({ video }) => {
 
     const { historyState, addToHistoryList, removeFromHistoryList } = useHistoryList();
 
-    const { playlistState, createPlaylist, addVideoToPlaylist, deleteVideoFromPlaylist } = usePlaylist();
-
     const { title, vidImage, creator, _id } = video;
 
     const likedItemExist = likesState.likedItems.find((vid) => vid._id === video._id);
@@ -31,9 +30,7 @@ const VideoCard = ({ video }) => {
 
     const isHistory = location.pathname === "/history";
 
-    const playlistHandler = () => {
-        createPlaylist(playlistName, video);
-    }
+    
 
     const modalHandler = () => setPlaylistModal(true);
 
@@ -112,84 +109,14 @@ const VideoCard = ({ video }) => {
 
                                     {/* PLAYLIST MODAL */}
 
-                                    <div className="modal_container flex flex_col flex_justify_center flex_align_center"
-                                        style={playlistModal === true ? { display: "flex" } : { display: "none" }} >
+                                    <PlaylistModal 
+                                        video={video} 
+                                        playlistModal={playlistModal} 
+                                        setPlaylistModal={setPlaylistModal} 
+                                        playlistName={playlistName} 
+                                        setPlaylistName={setPlaylistName} 
+                                    />
 
-                                        <div className="create_playlist_wrapper" >
-
-                                            <div className="modal_wrapper flex flex_col flex_justify_center">
-
-                                                <div className="modal_close_btn flex flex_justify_between">
-
-                                                    <h3>Playlist</h3>
-                                                    <i class="fa-solid fa-circle-xmark " onClick={() => setPlaylistModal(false)}></i>
-                                                </div>
-
-                                                {/* List all the playlists available in playListItems array */}
-
-                                                <div className="lists_playlist flex flex_col">
-
-                                                    {playlistState.playlistsItems?.map((eachPlaylist, i) => {
-
-                                                        // checking all videos in eachPlaylist with the given video if present or not
-
-                                                        return eachPlaylist.videos.find((vid) => vid._id === video._id)
-
-                                                            ? (
-                                                                <div className="playlist_list_btns flex flex_align_center">
-                                                                    <i
-                                                                        className="fa-solid playlist_one fa-check icon_circle_plus"
-                                                                        onClick={() =>
-                                                                            deleteVideoFromPlaylist(eachPlaylist._id, video._id)
-                                                                        }>
-                                                                    </i>
-                                                                    <span className="play_title">
-                                                                        {eachPlaylist.title}
-                                                                    </span>
-                                                                </div>
-
-                                                            ) : (
-                                                                <div className="playlist_list_btns flex flex_align_center">
-                                                                    <i
-                                                                        className="fa-solid fa-plus playlist_one icons icon_circle_plus"
-                                                                        key={eachPlaylist._id}
-                                                                        onClick={() =>
-                                                                            addVideoToPlaylist(eachPlaylist._id, video)
-                                                                        }>
-                                                                    </i>
-                                                                    <span className="play_title">{eachPlaylist.title}</span>
-                                                                </div>
-                                                            );
-                                                    })}
-                                                </div>
-
-                                                <div className="add_new_playlist">
-
-                                                    <div className="input_playlist_wrapper">
-                                                        <input
-                                                            type="text"
-                                                            className="input"
-                                                            placeholder="Enter Playlist Name"
-                                                            onChange={(e) =>
-                                                                setPlaylistName((prev) => ({
-                                                                    ...prev,
-                                                                    playlist: e.target.value,
-                                                                }))
-                                                            }
-                                                        />
-                                                    </div>
-
-                                                </div>
-                                                <button className="create_playlist_btn flex flex_justify_center" onClick={() =>
-                                                    playlistHandler()
-                                                }>
-                                                    <i class="fa-solid fa-circle-check icon_circle_plus">
-                                                        <span > Create New Playlist</span>
-                                                    </i>
-                                                </button>
-                                            </div>
-                                        </div>
-                                    </div>
                                 </div>
                             </div>
                         </div>
