+++ src/components/index.js
@@ -2,3 +2,4 @@ export {Header} from "./Header/Header";
 export {VideoCard} from "./VideoCard/VideoCard";
 export {PlaylistCard} from "./PlaylistCard/PlaylistCard";
 export {PlaylistVideoCard} from "./PlaylistVideoCard/PlaylistVideoCard";
+export {PlaylistModal} from "./PlaylistModal/PlaylistModal";
